+++ src/Is.php
@@ -20,15 +20,22 @@ final class Is implements Constraint
     private $assert;
     /** @var non-empty-string */
     private string $type;
+    /** @var ?non-empty-string */
+    private ?string $message;
 
     /**
      * @param pure-callable(T): bool $assert
      * @param non-empty-string $type
+     * @param ?non-empty-string $message
      */
-    private function __construct(callable $assert, string $type)
-    {
+    private function __construct(
+        callable $assert,
+        string $type,
+        ?string $message = null,
+    ) {
         $this->assert = $assert;
         $this->type = $type;
+        $this->message = $message;
     }
 
     public function __invoke(mixed $value): Validation
@@ -36,7 +43,9 @@ final class Is implements Constraint
         /** @var Validation<Failure, U> */
         return match (($this->assert)($value)) {
             true => Validation::success($value),
-            false => Validation::fail(Failure::of("Value is not of type {$this->type}")),
+            false => Validation::fail(Failure::of(
+                $this->message ?? "Value is not of type {$this->type}",
+            )),
         };
     }
 
@@ -151,6 +160,16 @@ final class Is implements Constraint
     public static function associativeArray(Constraint $key, Constraint $value): AssociativeArray
     {
         return AssociativeArray::of($key, $value);
+    }
+
+    /**
+     * @param non-empty-string $message
+     *
+     * @return self<T, U>
+     */
+    public function withFailure(string $message): self
+    {
+        return new self($this->assert, $this->type, $message);
     }
 
     public function and(Constraint $constraint): Constraint